[#47] Renaming ./models/index.ts to db.ts
