+++ package.json
@@ -2,7 +2,7 @@
   "name": "untis-to-calendar",
   "version": "1.0.0",
   "description": "",
-  "main": "src/index.ts",
+  "main": "src/index.tsx",
   "type": "module",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1 /*",
